refactor: use build pattern for creating enrolment filters
