--- src/app/routes/enrolment/enrolment-list/filters/filter-by-did/filter-by-did.ts
@@ -1,13 +1,0 @@
-import { EnrolmentClaim } from '../../../models/enrolment-claim.interface';
-
-export const filterByDid = (
-  list: EnrolmentClaim[],
-  value: string
-): EnrolmentClaim[] => {
-  if (!value) {
-    return list;
-  }
-  return list.filter(
-    (item) => item.subject.includes(value) || item.requester.includes(value)
-  );
-};
