--- src/app/routes/enrolment/enrolment-list/filters/filter-by-namespace/filter-by-namespace.spec.ts
@@ -1,22 +1,0 @@
-import { filterByNamespace } from './filter-by-namespace';
-
-describe('filterByNamespace function', () => {
-  it('should return the same list when filter value is falsy', () => {
-    expect(filterByNamespace([{} as any], '')).toEqual([{} as any]);
-  });
-
-  it('should remove 1 element from list', () => {
-    expect(
-      filterByNamespace(
-        [{ namespace: 'test' }, { namespace: 'cool' }] as any[],
-        'te'
-      )
-    ).toEqual([{ namespace: 'test' }] as any);
-  });
-
-  it('should not be case sensitive', () => {
-    expect(filterByNamespace([{ namespace: 'test' }] as any[], 'TE')).toEqual([
-      { namespace: 'test' },
-    ] as any);
-  });
-});
