--- src/app/routes/enrolment/enrolment-list/filters/status-filter/status-filter.spec.ts
@@ -1,49 +1,0 @@
-import { statusFilter } from './status-filter';
-import { EnrolmentClaim } from '../../../models/enrolment-claim.interface';
-import { FilterStatus } from '../../models/filter-status.enum';
-
-describe('statusFilter', () => {
-  const list = [
-    { isAccepted: false, isRejected: false },
-    { isAccepted: false, isRejected: true },
-    { isAccepted: true, isRejected: false },
-    { isAccepted: true, isRejected: false, isRevoked: true },
-  ] as EnrolmentClaim[];
-
-  it('should return all elements', () => {
-    expect(statusFilter(list, FilterStatus.All)).toEqual(list);
-  });
-
-  it('should return only rejected elements', () => {
-    expect(statusFilter(list, FilterStatus.Rejected)).toEqual([
-      {
-        isAccepted: false,
-        isRejected: true,
-      },
-    ] as EnrolmentClaim[]);
-  });
-
-  it('should return only approved elements', () => {
-    expect(statusFilter(list, FilterStatus.Approved)).toEqual([
-      {
-        isAccepted: true,
-        isRejected: false,
-      },
-    ] as EnrolmentClaim[]);
-  });
-
-  it('should return pending elements', () => {
-    expect(statusFilter(list, FilterStatus.Pending)).toEqual([
-      {
-        isAccepted: false,
-        isRejected: false,
-      },
-    ] as EnrolmentClaim[]);
-  });
-
-  it('should return only revoked elements', () => {
-    expect(statusFilter(list, FilterStatus.Revoked)).toEqual([
-      { isAccepted: true, isRejected: false, isRevoked: true },
-    ] as EnrolmentClaim[]);
-  });
-});
